fix(fiber): join handles panic if dropped unjoined
When the join handle get's dropped there's no way to join the fiber
anymore. But if it is left joinable at this point, the tarantool
internals won't be able to recycle it's memory. So we need to make sure
the join handle is joined explicitly at some point in the code.