feature(hlua): multiple return values
`LuaFunction::call` and `LuaFunction::call_with_args` have been updated
to support multiple return values, however they cannot check if the
expected number of values is the same as the actual one, so there is a
possibility of undefined behavior when using them for this purpose.

`LuaFunction::call_multi` has been added to solve the problem of
unsafety. It will check that the expected value count is the same as
the actual one

TODO: Update LuaRead trait to include information of expected number of
values on the stack