fix(build): fix build fail on rust 1.61
since 1.61 rust has added a `whole-archive` modifier for link targets
which is unset by default.

Previously to this version this was probably undefined behavior, because
the linking would work fine when building an executable (like picodata)
but would fail with undefined symbols when building a static library.

But now with the advent of this option, we can probably refactor
picodata into a library which can be useful for testing, i.e. we can
exclude testing stuff from the binary but still have a test binary with
all the necessary runtime symbols.

Closes #81