chore(picodata): picodata fork compatibility
adds "picodata" feature intended for enabling support
forTARANTOOL_EXECUTABLE=tarantool-pd  the picodata tarantool fork, which
is off by default. This allows us to both support vanila tarantool and
picodata/tarantool

`test-pd` target is added to `Makefile` which compiles
`tarantool-module-test-runner` with the `picodata` and runs the test.
The path to the `tarantool` executable can be specified via the
`TARANTOOL_EXECUTABLE` environment variable.