fix: don't use tarantool synchronous replication, it's broken
Tarantool synchronous replication is broken in such a way, that local
& asynchronous transactions get rolled back upon synchronous
transaction rollback. This is critical for us, because we use local
spaces for the raft log, which should be orthogonal to synchronous
business spaces, but is not due to this behavior. Until it is fixed, we
will not use synchronous replication.

See https://github.com/tarantool/tarantool/issues/7903