refactor: remove commit_index from Peer
Originally it was meant to be a key for distinguishing outdated peer
info. But this was never actually fully implemented. Now that we have
the governor and it's synchronizing mechanism there's even less point
to this field. And now that this legacy code impedes refactoring, we
remove it at least until we have a better idea of how to implement it.