fix(governor): don't bootstrap vshard until replication factor is satisfied
NOTE: there's still a bug, because we set each new replicaset's weight
to 1 (even ones which don't satisfy replication factor)
before the bucket distribution is bootstrapped. But there must be at
least one replicaset with non zero weight in order for vshard.*.cfg to
work.

A potential solution would be to only configure vshard once a replicaset
is filled up.