build: fix packpack
The build process was refactored recently in b9aeab59a and it broke the
packaging in a tricky way.

Packpack wraps the whole build process in a Makefile which exports its
own `MAKEFLAGS` environment variable:

```
MAKEFLAGS="w -j -- BUILDDIR=/build"
```

The `cmake` crate used to override it with `CARGO_MAKEFLAGS` provided by
cargo. But after we've replaced it with `std::process::Command`, the
`BUILDDIR` setting has made its way to the `icu` build, which suddenly
clashed one of internal `make` variables and wrote there resource files.
As a result, the `icu` build started to fail with the error:

```
icupkg: unable to open input file
"/build/usr/src/debug/picodata-22.11.0.158/
   /target/debug/build/tarantool-sys/icu-prefix/src/icu-build/
   /data/out/build/icudt62l/af.res"
```

while all resources were copied directly to `/build/af.res`.

This patch repeats the behavior of the `cmake` crate.

It also makes logging more verbose and aligns the output. Replacing
`eprintln` with `println` is necessary since `make` and `cmake` both
write the logs to the `stdout`.