fix: supress dead code stripping in the linker on macos
We don't want to strip unused symbols from the static libraries on
the linking stage. This leads to problems with tarantool's crypto
module. Some symbols ("crypto_EVP_get_digestbyname" for example)
were used only by the ffi calls inside `crypto.lua` file. So the
linker stripped these symbols as it was not able to find their
usage in the rest of the C code. This produced SIGABRT on the
picodata startup when tarantool_lua_init() initialized the crypto
module.