feat: use only primary index for ddl and dml
Related issues:
- #237 Restrict cas predicate over primary index only.
Having secondary index in the cas predicate is problematic. At the moment we don't store enough information to check the predicate and storing this information would make the logic much complex
and was deemed not important enough for the user.

- #258 Dml semantics
Even though tarantool supports delete/update by secondary key, in the write-ahead log the operations are stored by primary key only. It was decided to also proceed in this direction. The operations will be represented as by primary key only and any secondary key update logic will probably be offloaded to clients/interfaces.