feat: support index options of double type
Vinyl spaces have specific index options in the system "_index"
space that can be of the type "double". A good example are the
"run_size_ratio" or "bloom_fpr" options, that are always present
in the "_index" tuple for vinyl spaces. Previously, we failed to
decode such tuples.

Current commit extends the "Value" enum with a new "double"
variant. As far as we don't expect NaN or infinums in the tuples,
it seems safe to implement custom "Hash" and "Eq" traits for the
double values containing f64.