refactor: move cli related stuff to src/cli
main.rs has grown a bit, adding new stuff to it doesnt feel natural.
Smaller files are easier to digest. For me it is easier to find stuff
when there are less files on the top level and some hierarchy shows.
where to go. A while back I wanted to move sql command to separate
file but there is already sql.rs on the top level.

This patch solves this issue.