chore(ci): fetch picodata tags too, tweak test-helm-docker
When using git describe fetch depth for shallow clones needs to be deep
enough to reach exiting tag. We used to only deepen our submodules for
tarantool build but now we reached the same problem for picodata. So
fetch tags for it as well.

allow_failure in test-helm-docker-image makes CI properly fail and
prevent merges when there are failures. Without it it didnt go past
"manual action required" icon and allowed to merge CI without fixing
other test jobs