test: simplify test_audit.py
There was a huge amount of boilerplate code with redundant abstractions.
The test code must be as simple as possible and making abstractions over
simple data is not a way to go, especially when builtin types such as
dict do the trick just as easily.

The prettiness of the test code should be out lowest priority.