fix(pgproto): non-admin users can't read from query cache
This commit fixes that non-admin users can't read queries from the query cache.
An attempt to read from the cache used to result in the following error:
sbroad error: failed to get space_def: box error: AccessDenied: Read access to space '_pico_table' is denied for user 'alex3'

Previously, an error would occur when a user ran the same query twice.
During the first execution, the query was prepared and cached. However,
on the second execution, the client attempted to retrieve the query from the
cache and encountered an access denied error.