rework free port detection in tests
The patch changes how ports are assigned to nodes started in tests.
Previously port range was given to any xdist worker and each test chose
ports from the beginning of the range which can lead to conflicts. So
with the patch tests can no longer have interleaving port ranges.

Additionally ports were previously checked but the check was failing the
whole test in case of a failure. With the patch occupied ports are
skipped instead.

The patch also fixes a source of flakiness in PluginReflection
PluginReflection