fix: flaky sql tests failing with timeout
- Some sql queries that read from sharded tables
can fail with timeout error, but because of rethrows
it is showed as 'Connection reset error'. In order
to avoid retrying such tests in CI we can use retries
in tests. Note: solution to the problem itself is in
the progress, but we don't know when it will be done:
https://github.com/tarantool/vshard/pull/442

- Another solution is to use global tables instead of
sharded tables in tests where it does not matter.

- This commit uses both approaches: when possible sharded
tables were replaced with global ones. In other cases retries
were used.