refactor: move dropping database files from supervisor to child
This moves most of the logic from the supervisor process which helps
with a number of problems with the main_cord stuff. For example in a
following commit the context will only be parsed in tarantool::main_cb
which means config parameters aren't accessible in supervisor.

This also moves us towards getting rid of the supervisor process in the
future.