fix: remove redundant proc_sharding call when instance reincarantes
I'm not sure but I think it was added before we switched to counting
vshard config versions. Basically it was fixing the same A-B-A problem
but only in a single case.

Just goes to show how you shouldn't apply hacky fixes to problems you
don't fully understand.

BTW nobody is ever reading this message, the commit only removes the
code so nobody will ever bump into it going over git blame...