fix(test): granting privileges to nonexistent user
This commit restores the user name in the ALTER USER command to "Dave" after
it was mistakenly changed to "DAVE", a nonexistent user. This change
broke the test logic, causing the test to erroneously verify granting
privileges to a nonexistent user, rather than confirming that we can
grant default privileges. The previous implementation resulted in noop,
so this went unnoticed until we started validating user existence.

Here is the commit changing Dave to DAVE:
https://git.picodata.io/picodata/picodata/picodata/-/commit/ecbb520b559ceb7f2dbe7ae0462c8d062faf1589