test: cleanup cluster.cas vs instance.cas inconsistencies
Note that now we don't have any (is this right?) tests that use pico.cas
which should be considering we plan to get rid of it in the near future.

Also we don't use `password` for cas operations, because the stored
procedure can only be called by the superuser and only the access
control checks rely on the `as_user` parameter.