fix(http_server): consider timeout for establishin TCP connections in HTTP server
Due to the bug in RPC (see https://git.picodata.io/picodata/picodata/picodata/-/issues/991)
RPC call timeout does not account TCP connection establishment.
So a request to a peer without a connection was not aborted
due to timeout.
Force a timeout directly for a future returned by RPC to avoid this

Closes #991