feat(plugins): add picodata version compatibility check
This commit embeds in plugins a new variable, containing picoplugin
version used to build it. If this version differs from the picodata version
when the plugin gets loaded, the loading fails with an error reporting
the expected and the current versions. This check can be disabled via env
PICODATA_UNSAFE_DISABLE_PLUGIN_COMPATIBILITY_CHECK.