fix: don't update current_master_name until it synchronizes
This patch changes when we update current_master_name during consistent
master switchover. Now it is updated as soon as we determine that the
new master has synchronized it's vclock with the old master.

As a result we can now block expel of a non-last replica in cases where
other replicas are offline, which is crucial because instances can
easily become offline due to temporary network issues.