build: tune release profile
As shown in
https://git.picodata.io/picodata/picodata/picodata/-/issues/1026#note_107222,
enabling LTO and using only one codegen unit can accelerate some
scenarios by 16.89%.

This commit adds these options to the "release" profile. However, it makes
the build much slower, so a new profile "fast-release" is added to address
this problem in CI. The "release" profile will only be used when merging
to master.