fix: flaky test_extreme_integer_values
This test is flaky due to vhsard rebalancing, so migrating to global
tables resolves the issue, since we don't use vshard for them. However,
after this migration one of the `pytest.raises` blocks's been failed,
as turned out, because u64::max value cannot be inserted in global tables.
We decided to don't fix this problem in this commit, so this check
was removed from the test and a new issue was added.

New issue: https://git.picodata.io/core/picodata/-/issues/1216