test: setup logging in tests
Logs help to debug flaky tests. Because sometimes just by instance logs
it is not that easy to guess what the test was doing at that time.
Especially when there is a repetitive pattern in the test.

For now instance logs and test output are printed in separate sections.
It still allows to match events by time easier compared to looking only
at instance log (it is way bigger compared to test log).

Additionally this patch sets up a `framework` package intended to host
reusable test code without polluting conftest.py